+++ spec/models/hierarchy_spec.rb
@@ -14,13 +14,13 @@ describe Monarchy::Hierarchy, type: :model do
     let!(:memo1) { create :memo, parent: project }
     let!(:memo3) { create :memo, parent: project }
 
-    subject { described_class.in(project) }
+    subject { described_class.in(project.hierarchy) }
 
     context 'when model is not monarchy resource' do
       let!(:user) { create(:user) }
 
-      it { expect { described_class.in(user) }.to raise_exception(Monarchy::Exceptions::ModelNotResource) }
-      it { expect { described_class.in(nil) }.to raise_exception(Monarchy::Exceptions::ResourceIsNil) }
+      it { expect { described_class.in(user) }.to raise_exception(Monarchy::Exceptions::ModelNotHierarchy) }
+      it { expect { described_class.in(nil) }.to raise_exception(Monarchy::Exceptions::HierarchyIsNil) }
     end
 
     it do
@@ -140,8 +140,9 @@ describe Monarchy::Hierarchy, type: :model do
 
     context 'accessible_for in' do
       let!(:memo_member) { create(:member, user: user, hierarchy: memo4.hierarchy) }
+      let(:hierarchy) { memo2.hierarchy }
 
-      it { expect(described_class.accessible_for(user).in(memo2)).to match_array([memo3.hierarchy, memo4.hierarchy]) }
+      it { expect(described_class.accessible_for(user).in(hierarchy)).to match_array([memo3.hierarchy, memo4.hierarchy]) }
     end
   end
 end